# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "patch": {
        "operationId": "UsersController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deactivateUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/password": {
      "patch": {
        "operationId": "UsersController_updateUserPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/sales": {
      "get": {
        "operationId": "SalesController_listSales",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sales"
        ]
      },
      "post": {
        "operationId": "SalesController_createSale",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Sales"
        ]
      }
    },
...
    "/sales/{id}/send-whatsapp": {
      "post": {
        "operationId": "SalesController_sendTicketWhatsApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendSaleWhatsAppDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Sales"
        ]
      }
    },
...
    "/reservations": {
      "get": {
        "operationId": "ReservationsController_listReservations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      },
      "post": {
        "operationId": "ReservationsController_createReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{id}": {
      "get": {
        "operationId": "ReservationsController_getReservation",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      },
      "patch": {
        "operationId": "ReservationsController_updateReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{id}/cancel": {
      "post": {
        "operationId": "ReservationsController_cancelReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{id}/reassign": {
      "post": {
        "operationId": "ReservationsController_reassignReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{id}/payment": {
      "post": {
        "operationId": "ReservationsController_recordPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    },
...
    "/public/reservations/{token}": {
      "get": {
        "operationId": "PublicReservationsController_getByToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicReservations"
        ]
      },
      "patch": {
        "operationId": "PublicReservationsController_updateByToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePublicReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicReservations"
        ]
      }
    }
  },
  "info": {
    "title": "MAGIC CITY API",
    "description": "API demo funcional para manejo de eventos",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc_session"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserPasswordDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateSaleDto": {
        "type": "object",
        "properties": {}
      },
      "SendSaleWhatsAppDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateReservationDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateReservationDto": {
        "type": "object",
        "properties": {}
      },
      "CancelReservationDto": {
        "type": "object",
        "properties": {}
      },
      "ReassignReservationDto": {
        "type": "object",
        "properties": {}
      },
      "RecordPaymentDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdatePublicReservationDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}